Refactored `compilation_stage.py`
The `run_manual_compile()` method of the CompilationStage class was
doing a lot, and it felt beneficial to break it down further with
helper functions. No functionality or behavior was modified, and
behavior on test cases matches behavior of previous version.

Also, began work on the StaticAnalysis class (file was generated,
constructor was written)