Implemented a primitive, rigid Static Analysis Stage
Includes various refactors in the CompilationStage class.

Static Analysis works (and has been tested) locally on a Mac where `clang-tidy` has been given a symlink, such that the command `clang-tidy` can be called directly in terminal. Instructions on how to do this will be added to README at a later point.